+++ raffle/mixins/metaclass.py
@@ -1,0 +1,5 @@
+from abc import ABC
+from redbot.core.commands import Cog
+
+class MetaClass(type(ABC), type(Cog)):
+    pass
